// LinkDev.MOA.POC.Portal.Fortnend/ClientApp/src/modules/shared/Enums/lookup-request-mode.enum.ts
export enum LookupRequestMode {
    LookupWithId = 1,
    OptionSet = 2,
    LookupWithCode = 3,
    Product = 4,
    ContractSubmissionCRs = 5,
    Account = 6,
    ContactDetails = 7,
    KSACity = 8,
    LookupWithName = 9,
    ContractSubmissionContacts = 10,
    WorkspaceCRs = 11,
    WorkspaceContracts = 12,
    WorkspaceServiceDefinitions = 13,
    ServiceDefinitionDetails = 14,
    WorkspaceRequestStatus = 15,
    WorkspaceSpecialContractsTypes = 16,
    CRsByTypeOffice = 17,
    CRsByTypeContractor = 18,
    CRsByTypeInvestor = 19,
    ContractMainType = 20,
    FilteredOffice = 21,
    FilteredContractor = 22,
    PermitsType = 23,
    PermitRequestContracts = 24,
    PermitRequestSpecialContractsTypes = 25,
    Permits = 26,
    RequestDestination =27 ,
    Requester =28,
    CoordinationLetterContracts = 29,
    ContractorOldCertificate = 30,
    ContractorRequestTypes = 31,
    Projects = 32,
    ContractOptionModel=33,
    ModonProducts=34,
    ContractorOldRatingCategory = 35,
    WorkspaceTicketList = 100
}

	